Remove error when converting Finally field to alpha1
Prior to this commit pipelines returned an error when converting from
a v1beta1 pipeline to a v1alpha1 pipeline if that pipeline had a Finally
section in it. As a result of this error the kubeapi would repeatedly
request v1alpha1 pipelines every second, filling up our webhook logs
with these errors.

This commit removes the error when downgrading a beta1 resource to alpha1.
I actually have no idea if this is the correct thing to do and I can't find
any docs explaining the dos and donts of conversion. This stops the spam
though.