Short term fix for #2676
SelfLink is deprecated and is causing an error on Kubernetes v1.20
clusters as outlined in the issue. This is a short term fix that
would set the source field to a value that should match the current
source URI without the value of the auto-populated selfLink field.

Another field could be used for the source field without issue, but
could cause concerns about backwards compatibility.