Serialize and deserialize Finally in PipelineRuns too
In commit 3ea5981 we added a way for Pipeline.Spec.Finally to be serialized into
alpha Pipelines when downgrading from v1beta1. This was done to fix an
error occurring during conversion and to avoid potential data loss.

This commit adds a similar feature to PipelineRuns. Specifically, if a
PipelineRun nests a PipelineSpec with Finally then that too needs to be
serialized and stored on the alpha version to be deserialized later.