Serialize entire v1beta1 PipelineSpec when converting to v1alpha
Prior to this commit our conversion of pipelines from v1beta1 to v1alpha1
did not account for fields that were solely present in the v1beta1
pipeline spec. This meant, for example, that When Expressions in PipelineTasks
were quietly dropped when moving from v1beta1 to v1alpha1. If a user
subsequently re-applied that v1alpha1 resource to the cluster then
the resource would be upgraded back to v1beta1 prior to storage without
the missing when expressions. This problem would only be exacerbated further
over time - any new features added to the v1beta1 spec would be lost on
conversion.

This commit serializes the entire v1beta1 pipeline spec into an annotation
on Pipelines and PipelineRuns (for nested pipelineSpec fields). This will
protect the entire set of beta-only fields from being lost in translation.

Unfortunately this approach does also have drawbacks: if a user makes changes
to the v1alpha1 YAML and then re-applies it then those changes
will be lost - overwritten by the contents of the annotation. I'm not sure if
there's a great way to determine, for all fields, how to successfully "merge"
the beta annotation with the contents of the modified v1alpha1 spec when
up-converting.

An e2e test has been added to send a v1beta1 pipeline through
the entire roundtrip flow and then compare the before/after for
differences.

One more fix is also included here: prior to this commit we were not preserving
pipeline results when converting pipelines from v1beta1 to v1alpha1.
This commit corrects that by copying pipeline results to the v1alpha1 pipeline
and back.