Run conformance tests regularly 🏃‍♀️
In https://github.com/tektoncd/pipeline/pull/3400 we added an initial
conformance test, and added a build tag `conformance` to make it so that
folks could run just conformance tests if they wanted to. From what I
can see it doesn't look like we use this tag at all in our automation,
so I want to add it to the tags we use to run our end to end test to
make sure this continues working over time.

In the long run I think we might want to explore a couple other
alternatives for implementing conformance tests, e.g. including making
use of our example tests (though those dont express the desired end
state), maybe a tool like [kuttl](https://github.com/kudobuilder/kuttl#kuttl)
which @chhsia0 has mentioned in preivous working groups, or like the
approach that knative seems to be taking via
https://github.com/knative-sandbox/reconciler-test

But in the short run let's at least make sure this test keeps working :D

Also renamed the test to make it a bit more clear what this test is for
in case it fails.