Add doc links for 0.23.0 to README and update release-cheat-sheet
0.23.0 has just been released so this commit adds an entry to our README.md
for this release's docs and examples.

During the release process I documented a `curl`+`jq` command that the
release captain can use to get a complete list of all release names
used so far by Pipelines. This helps when checking for name collisions
while picking the next "<cat breed> <famous robot>" pair.

Also indents a section in the cheat sheet that was using 3 spaces to match
the 4-space indent used elsewhere in the same doc.

Also updates the release cheat sheet to mention the release in the #announcements
channel.