Updates the Tekton Bundle spec and resolver to be more explicit.
The current spec is ambiguous with regards to the form of the `kind` annotation. In the code, we implicitely enforce that the kind must be singular, and lowercased. The spec half-way asks for this and the resolver doesn't enforce this. Therefore, this change:

- adds validation to the oci resolver to enforce the annotation rules
- adds corresponding tests to ensure the enforcement is invoked
- modifies the bundle contract to explicitely state the pluralization requirement
- modifies a test utility for invoking this case

The tkn cli has been updated as well to conform to this explicit spec.