Make entrypoint binary build for Windows
The entrypoint binary includes non-portable code that only works on
Linux. This change adds runner_windows.go with a build constraint so
that it's only built when GOOS=windows, which removes this code (and the
functionality it enabled).

You can test this with:

KO_DEFAULTBASEIMAGE=hello-world ko publish ./cmd/entrypoint/ --platform=all

This tells ko to build the entrypoint binary using the hello-world image
as a base. The hello-world image doesn't really matter, but it includes
lots of platforms, including both Linux and Windows OSes.

Eventually we'll want to identify (or build) a better base image, this
is just to demonstrate that building the image works when built for
Windows.