Remove random suffixes on creds-init volumes
Removed the random suffixes that used to appear on creds-init volumes. These
didn't serve any purpose that couldn't be equally well-served with an
incrementing integer.

The benefit of doing it this way is that our unit tests become less brittle -
those random suffixes were a source of considerable churn whenever other
volumes were added or removed to test data because the random pool of suffixes
was shared with those other volumes.