Enable Bundles and Custom Tasks when alpha feature gate is set to "alpha"
Prior to this commit the Tekton Bundles and Custom Task features were only enabled
when their individual flags were set to "true" in the feature-flags configmap.

After this commit the two features are enabled if enable-api-fields is "alpha" or
if their individual flags are set to "true". Note that it is _not_ possible to
set the "alpha" feature gate but disable Bundles / Custom Tasks. These features
are "alpha features" so they're always enabled regardless of their own feature
flags if the gate is set to "alpha".