Add a nodeAffinity rule to the controller and webhook deployments
This commit prevents controller and webhook pods from being
scheduled on Windows nodes. Without these affinity rules, the
controller or webhook will fail to deploy to a cluster if they
are scheduled on a Windows node by Kubernetes.

There is no need for these two components to run on Windows nodes
for Tekton to support mixed Linux + Windows clusters.