Test multiple scripts in one task executing with different UIDs
Prior to this commit we weren't testing whether scripts were executable
when UIDs differed from Step to Step in a single Task.

This commit updates the step-script examples to vary the UID of each
Step. A failure of one of these would indicate that a bug has
been introduced in the permissions of the shell scripts that Tekton
generates and places in the /tekton/scripts directory.