Run integration tests in "alpha" as part of pipelines' CI
Prior to this commit the integration tests run against PRs
were only those for "stable" features.

This PR adds additional integration test runs for "alpha"
features as well. `test/e2e-scripts.sh` now patches the feature-flags
ConfigMap before executing integration tests and examples tests, first
with "enable-api-fields: stable" and then with "enable-api-fields: alpha".

As part of switching on alpha tests two issues have also been fixed
with alpha features:

1. Sidecar workspaces could collide with existing volumeMounts attached
to the sidecar if the mountpath of the workspace matched an existing
volumeMount. This has been fixed by preventing the additional workspace
volumeMount if one already exists at that mountpath.

2. Tekton bundles integration tests needed to be updated to account
for validation errors that are returned when a bundle is invalid.