Disable gocritic in pipelines
The test-runner docker image from plumbing now includes a newer version
of golangci-lint (1.39.0) which apparently switched on or changed some gocritic
configuration *somewhere*.

There are a _lot_ of syntax issues in pipelines that gocritic does not like,
including many in the generated code of `pkg/client`. Rather than hand-edit
everything or figure out all the flags required to disable the features this
commit simply disables gocritic so that PRs can get moving again.

We can re-enable it once we figure out what the necessary flags are to get
it working (or we fix all the syntax issues).