Update tektoncd/plumbing.
This is pulls in tektoncd/plumbing@3c5af40 to fix integration tests.

(cherry picked from commit 55ae8563be29a76bb43b9577c8d6072ae7110d78)

We need to cherry pick this change into v0.24 to run integration tests
for a Workspaces in Sidecars fix. We need to fix the Workspaces in
Sidecars to be serialized as workspaces not Workspaces. This is change
is made in https://github.com/tektoncd/pipeline/pull/4008. Its
integration tests are failing because basic authentication was removed
for GKE cluster versions >= 1.19 so clusters cannot be created with
basic authentication enabled. So we need to cherry pick this change to
disable basic auth and create clusters to run the tests for the
Workspaces in Sidecars fix.