Document the safe-to-evict annotation on the webhook deployment
The annotation is set by default because by default we assume Tekton
will be installed in a small cluster with only one replica of the
webhook deployment, and that replica should not be disturbed.

If an operator configures multiple replicas they can remove the
annotation to give a Cluster Autoscaler more freedom to scale down
nodes.