Add ConfigMap which can contain pipelines info
As of now we fetch version of pipelines through labels present on the
deployments which is read by tools such as `tkn cli` and display the
version. This version may not be displayed to users if they don't have
permission to view the deployment.

In this commit we are adding
1. A `ConfigMap` which contains version information.
2. RBAC which will give appropriate permissions to view the ConfigMap
irrespective of whether user is has permission to view other objects in
that namespace or not.

Signed-off-by: vinamra28 <vinjain@redhat.com>