Only fetch the definitions once 🧙
Since we "dereference" fetched definition and store them into the
`status`, we can use this as a source of truth. This allows us to
fetch only once the definitions (Task, Pipeline) *and* has the benefit
to make the runs (PipelineRun, TaskRun) immune to change to what they
refer, while the are executing.

Without this change, if you run a PipelineRun that reference a
Pipeline that is deleted during execution, your PipelineRun will fail
at some point because it cannot fetch the definition anymore, even if
it stored them in its status. This is the case for TaskRun too.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>