Fix Windows image entrypoint paths for parity with Linux
Tekton expects to find the entrypoint binary in '/ko-app/entrypoint'
which is equivalent to 'C:\ko-app\entrypoint' on Windows. See
'pkg/pod/entrypoint.go:orderContainers()'. Currently entrypoint is
stored in 'C:\ProgramData\tektoncd\pipeline\entrypoint.exe' causing
this to break on Windows. To avoid the need for platform-specific
logic in the controller, and for parity with Linux, this change
updates 'entrypoint' and 'nop' images for Windows to store binaries
in 'C:\ko-app\'.