Use tektoncd/results repo for git-init symbolic ref tests
Prior to this commit we were pointing one of our integration tests
at a Spring project repo. We did this to test our git-init code
with a repo that uses "main" as its main branch instead of "master".

Now that we have switched many/all of our repos over to "main"
instead of "master" we can use them for testing instead. This means
we haven't tied our tests to an external third party repo that might
change / be removed without our realizing it.