Set fake AWS credentials on controller to workaround aws-sdk bug
Several issues have now reared their head which are directly caused
by an update to the aws-sdk. The update results in extremely long
delays in the execution of tasks after the Pipelines controller is
first deployed in a cluster. The aws-sdk is initialized through
a transitive dependency that pipelines pulls in via k8schain.

Here are the recent issues directly related to this aws-sdk bug:

- https://github.com/tektoncd/pipeline/issues/3627 (Since December!)
- https://github.com/tektoncd/pipeline/issues/4084

One quick way to work around this problem is to set fake AWS
credentials in the environment of the deployed controller. This
apparently causes the aws-sdk to skip whatever process it has
introduced that causes massive delays in execution. So this commit
does exactly that - set fake aws creds in the deployments env vars.

This is an unfortunate hack to try and mitigate the problem until
a better solution presents itself.