specifying onError for a step
this commit implements tep-0049 - ignore a step error

When a `step` in a `task` results in a failure, the rest of the steps in the
`task` are skipped and the `taskRun` is declared a failure. If you would like
to ignore such step errors and continue executing the rest of the steps in
the task, you can specify `onError` for such a `step`.

`onError` can be set to either `continue` or `fail` as part of the
step definition. If `onError` is set to `continue`, the entrypoint sets the
original failed exit code of the script in the container terminated state.
A `step` with `onError` set to `continue` does not fail the `taskRun` and
continues executing the rest of the steps in a task.

This is an alpha feature. The `enable-api-fields` feature flag must be set to
`"alpha"` to specify `onError` for a `step`.

This commit includes following changes:

* Changing entrypoint to include three new flags `onError`, `stepMetadataDir`, and
`stepMetadataDirLink`.
* Adding one new function as part of the runner CreateDirWithSymlink
* Creating a volume `/tekton/steps/`
* Supporting a path variable $(steps.step-<stepName>.exitCode.path) and
$(steps.step-unnamed-<stepIndex>.exitCode.path)
* API spec `onError` while defining a step
* Writing exitCode at /tekton/steps/step-<step-name>/exitCode or
/tekton/steps/step-unnamed-<step-index>/exitCode
* Set the exitCode of a terminated state to a non-zero exit code
* Doc, unit test, and examples for this feature