Add issue link to comment in controller yaml
In commit 49a7fa2db63fc633190da260d7167c05a8a7a91d we added
a workaround for an issue that can cause the tekton controller
to be delayed launching taskruns for the first time.

This commit updates the comment on the workaround to include
an issue link with more information about the reason for the
workaround.