Always refer to `finally` tasks 📖
This commit changes all references to "final tasks" and "finally tasks"
to `finally` tasks".  We were using a mix of "final tasks" and "finally
tasks" to refer to these tasks, so this commit is making us consistently
use one version.
"finally tasks" is less grammatically correct that "final tasks"
because finally is an adverb - "finally tasks" only makes sense if
"finally tasks" is considered short for "finally tasks in a pipeline".
We're working around this by using code block syntax to indicate that
`finally` is a symbol in the Pipeline spec, and not just the adverb
"finally".

Another option sould be to go the other way and always use "final tasks"
(and never say "finally tasks"). However if we do this, @Srivaralakshmi
pointed out in https://github.com/tektoncd/pipeline/issues/4086 that
this may make it confusing in that reading something like 'the final tasks'
may make a user expect to see a seciton in the pipeline called `final`
instead of the actual section `finally`.

We have another option which is to consider `finally`- an adverb - a
bad name for that section of the Pipeline spec and change it to `final`.
My understanding of the syntax is a bit different, e.g.:
```yaml
spec:
  tasks:
    - name: tests
      taskRef:
        name: integration-test
  final:
    - name: cleanup-test
      taskRef:
        name: cleanup
```
However that would be a big change (tho pre-v1 would be
the time to do it!) and `finally` is consistent with how this kind of
functionality is expressed in languages like Python.
(Also I understand the finally section to be saying: once all of the above is
done, finally run clean-test, and not to say "cleanup-test is a finally
kind of task".)

The last option would be to support both final and finally but that
feels to me like it would be even more confusing.