Loosen permission checks on TestCp to handle umasks.
System umasks can futher restrict permissions passed to os.OpenFile,
resulting in a file that's more restricted than the permission value we
supplied. Instead of trying to match the permissions exactly, this
change modifies the check to make sure that the new file does not exceed
the expected permissions, but allows anything more restrictive.