Take advantage of `kmeta.OwnerRefable`.
I noticed that the tekton `controller.go` files were inconsistently using `controller.HandleAll` (cosmetic), but also that several of the `impl.EnqueueControllerOf` lines in the PipelineRun controller weren't filtering on the kind of the owner (can cause unnecessary reconciliation).

For the latter, we also introduced a nice new helper `controller.FilterController(&v1beta1.TaskRun)` that takes advantage of `kmeta.OwnerRefable` to filter on the GVK in the owner reference.

This spiraled a bit because it turns out Tekton wasn't taking advantage of `kmeta.OwnerRefable` or `kmeta.NewControllerRef`!  So the majority of the churn in this change is:
 - `GetOwnerReference` -> `GetGroupVersionKind` (implements `kmeta.OwnerRefable`)
 - `foo.GetOwnerReference()` -> `*kmeta.NewControllerRef(foo)`

As some background, our `kmeta.NewControllerRef` API was modeled after this from API machinery: https://pkg.go.dev/k8s.io/apimachinery/pkg/apis/meta/v1#NewControllerRef

We added our own variant to drop the tedious GVK requirement, since the first argument should be sufficient (thus `kmeta.OwnerRefable`).

/kind cleanup