Drop the tracker.
Tekton's use of the `tracker` is redundant with the informer events it already has configured, and the tracker lease the pipeline run controller is using is probably too small to actually be effective for long-running pipelines.

The tracker is built around the premise that resyncs happen to "refresh" things, and so to be safe the tracker lease should be a multiple of the resync period.  The pipeline controller is currently hardcoding `30m`, but the default resync period is `10h`, so anything taking longer than `30m` via the tracker will actually drop events.

Fortunately, I believe that both usages of the tracker are wholly redundant with informer events on owned resources anyhow, so this change should just remove unneeded complexity.

For additional background, the purpose of the tracked isn't to track owned resources, but rather to track resources referenced via `corev1.ObjectReference` (or equivalent).  We use this in Knative to track changes to Addressable resources that some resource might be delivering events to (e.g. maybe the address changed, or it is no longer Ready).