Add script mode support to windows tasks
Steps and sidecars can contain a script field. In linux tasks these
scripts are copied into files which are made executable and then
steps are added to the task to execute those files. This commit adds
comparable functionality if scripts are used in a task which will run
on a windows node.

On a windows node the mechanics are different, due to how windows
handles executable files. The key difference is that Tekton needs to
know that a script will run on windows, and how to run the file
(which interpreter to use). This is done through a ‘windows shebang’
line at the start of the script.

The line must begin with ‘#!win’. After that the user needs to provide
the interpreter to use, as well as any necessary arguments. The line
must be written such that the name of the file containing the script to
execute can be appended to the end.

For example, to run the script in the file ‘test.ps1’ with powershell,
the command would usually be ‘powershell -File test.ps1’ and so the
shebang line must be ‘#!win powershell -File’.

If no interpreter is provided (i.e. the shebang line is only ‘#!win’) then
the script contents will be stored in a .cmd file and executed.

Finally, since a pod cannot contain a mix of windows and linux
containers a windows shell image has been added to the Images
structure, which will be used in the place-scripts step when needed
on a windows node.

To maintain parity with other alpha features, task validation for tasks
containing windows scripts will now require the 'enable-api-fields' flag
to be 'alpha'. TaskRuns/Tasks that contain windows scripts will be
rejected if this flag is not set, giving the user immediate feedback.

The integration tests for windows scripts have been updated to reflect
this alpha flag requirement.