Cancelling a Task with Retries - recognize as Failure
`Finally Tasks` are executed only when all `Tasks` are _Done_, where
_Done_ means a `Task` is _Skipped_, _Succeeded_ or _Failed_. But a
`Task` is considered _Failed_ only when its `Retries` are all done.

Today, when a `Task` with `Retries` is `Cancelled`, it's not considered
_Failed_ if its `Retries` are not yet done. This is problematic because
if a `PipelineRun` is _Gracefully Cancelled_, where it should run
`Finally Tasks`, and it has a running `Task` with `Retries`:
- the `Task` with `Retries` be cancelled, but won't be considered
_Failed_ because its `Retries` are not done
- this would block the execution of `Finally Tasks`
- the `PipelineRun` would hang (stay running) until it times out

In this change, we modify `Is_Failure` to consider that either all
`Retries` are done or `Reason` for failure is `Cancellation`. With this
change, we can unblock execution of the `Finally Tasks` during _Graceful
Termination_.

In addition, the documentation notes that "`Retries` - Specifies the
number of times to retry the execution of a Task after a failure.
_Does not apply to execution cancellations_."

Fixes https://github.com/tektoncd/pipeline/issues/4125