Fix Example with Completed and Succeeded Execution Status
Today, an example `pipeline` that demonstrates using the execution
status of `tasks` - `$(tasks.status)` - in `when` expressions in
`finally tasks` uses `Succeeded` in a `Task` that should be executed
but the status is actually `Completed`. This example would be confusing
to users as reported on Slack.

As explained in the [docs](https://github.com/tektoncd/pipeline/blob/release-v0.24.x/docs/pipelines.md#using-aggregate-execution-status-of-all-tasks):
- `Succeeded`: all `tasks` have succeeded
- `Failed`: one ore more `tasks` failed
- `Completed`: all `tasks` completed successfully including one or more
skipped `tasks`

In this change:
- correct the `task` that should be executed to check `Completed`
instead of `Succeeded`
- add a `finally task` that shows `Succeeded` would block execution
because the aggregate execution status is actually `Completed`