Try reducing "Invalid log format" errors
Over in another PR (github.com/tektoncd/pipeline/pull/4108) I was seeing a lot
of "Invalid log format for pod" errors that looked to be coming from logstream.

After chatting with mattmoor it looks like the values we've got in our
config-logging need to be updated. This PR matches those values to knative
serving to see if that potentially reduces the messages.