Change "fail" to "stopAndFail" 🧐
This commit proposes changes the wording of the `fail` option for
`onError` from `fail` to make it more explicitly clear that this option
(which is the default behavior when `onError` is not specified) will
both cause the Task to fail AND stop execution of any subsequent steps.

This change doesn't propose changing any of the described functionality,
just the value of the field.

Change to the TEP should be approved before this is merged
(https://github.com/tektoncd/community/pull/497) and if we go ahead with
it we should do a patch release ASAP to minimize user impact - that
being said, this change impacts the default value of the field which
users are unlikely to specify explicitly - users are much more likely to
start using `onError: continue` and for them there would be no impact.