cleaning up label keys
We have defined GroupName as a constant which is set to tekon.dev in
register.go. We have other label keys defined in the same go file which
are used by the controller to set the appropriate labels. These keys
are TaskLabelKey, PipelineLabelKey, etc. These keys are set to
"/task", "/pipeline", etc. Now while adding such labels in the tekton
resources, we join GroupName to the key. Instead, this commit changes the way
we have defined a label key to include GroupName as part of the label key and
use just the label key instead of concatinating while using the key. This
cleansup the code and makes it simple to add tekton internal labels and
retrieve them.