Patch pipeline name label with GroupName to avoid back compat change in 0.27
The pipeline name label propagation change in 2a910a3 was merged after a slight
change in the way we construct some publicly exported label constants. Bringing
in the change to public constants would have technically been breaking for a
patch release.

This reintroduces GroupName to the label construction when propagating the
pipeline name label to workaround the conflict.