Deduplicate timeout constants.
Trying to set up a "downstream test" for Tekton (to detect when knative.dev/pkg changed break Tekton), I noticed that Tekton had a duplicate constant `pipelineRunTimeout`, which is a problem when a build is run with both Go tags).

When I went to combine/move this into `wait.go` (no build tag) I noticed there was already a `timeout` constant with an identical value.  So I simply removed both constants and changed the uses of the old constant to `timeout`.

With this change, and another to our downstream-test action, I will hopefully be able to add the check I want on knative.dev/pkg.

/kind cleanup