Improve the logging in pkg/git testing
Currently the "fetch" testing nicely captures logs (to examine output), but doesn't actually surface them on failures.

This change adds a `defer`red function that stream the output to `t.Log` so that `git` failures can be debugged.

For example, when I introduce a typo into the "commit" command, it now prints:

```
=== RUN   TestFetch/test-clone-with-sparse-checkout
    git_test.go:299: exit status 1
    git_test.go:216: ["error" git]: Error running git [codmmit --allow-empty -m Hello Moto]: exit status 1
        git: 'codmmit' is not a git command. See 'git --help'.

        The most similar command is
                commit
```

Previously, this would only print the exit status, which isn't particularly useful.