Un-randomize TestConvertScripts_WithBreakpoint_OnFailure
Prior to this commit we iterated over a map when building up the args
passed to a step container with debug breakpoint enabled. Map iteration
can be random, so sometimes our unit tests would flake out.

This commit switches the map to a list and uses a struct to provide
the fields as the key/values from the map.