Converts ResultType in PipelineResourceResult string to int
This converts the ResultType in PipelineResourceResult from string
to int enum to curtail the size of json object emittted by steps.
ResultTypes were removed because they made JSON messages bigger, which in
turn limited the amount of space in termination messages for task results. String
support is maintained for backwards compatibility.

Closes #4150

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>