Bump the go-scm, docker/cli and docker/docker libraries.
Both of these were quite out of date and have enough transitive dependencies
that it's worth keeping closer to HEAD. It might be possible to detangle and
update separately, but I couldn't get it to work on my first try.

There was one manual change to error strings in pkg/pullrequest/api_test.go.
I also had to re-run codegen and update the golden test data for new fields..

Signed-off-by: Dan Lorenc <dlorenc@google.com>