Bump golangci-lint to v1.42.0 from v1.30.0.
Also fix the resulting lint errors, and a few of the resulting lint warnings.
One of the main linters we use (golint) is now deprecated. The replacement is known
as revive, and golangci-lint displays a warning about this now.

Unfortunately quite a few lint warnings appear during that update, so I'm doing it in a few pieces.

Signed-off-by: Dan Lorenc <lorenc.d@gmail.com>