Remove the pkg/jsonpath library.
This was originally added as part of #1951 and #1493 in January 2020,
but the work was never finished and the package isn't used anywhere.

It's a non-trivial amount of code and maintenance to keep up to date as we
bump linters and scanners, so we should delete it until there's a clear use
case.

I'm not aware of anyone using this now, so another reason to remove it now is to
prevent someone from taking a dependency on it.

Signed-off-by: Dan Lorenc <dlorenc@google.com>