Make `NewRecorder` a proper singleton
`NewRecorder` cannot be called twice safely because it would result in the same view being registered twice, which always returns an error.

The tests work around this by themselves manually unregistering the metrics that `NewRecorder` registers, so an alternative here might be to have `NewRecorder` also return a `context.CancelFunc` to unregister things, or to have `NewRecorder` take `ctx` so that it can unregister the metrics on context-cancellation.

This change simply opts for making `NewRecorder` a proper singleton with a `sync.Once` and private globals for the `Recorder` and the possible error, so multiple calls to `NewRecorder` return the same singleton `Recorder` (or its error).

I've updated the tests to clear this new singleton state (as well as continuing to manually unregistering the metrics).