Poll for the status of Runs
Today, each custom tasks implementation copies over the wait.go file
to poll the status of resources for testing. This makes it hard to
maintain over time. The only added function in the custom tasks testing
is the polling for the status of Runs.

This change adds the function to poll for the status of Runs so that
this can be reused across all custom tasks testing.