Bump github.com/cloudevents/sdk-go to v2.5.0
This updates the cloudevents dependency from v2.1.0 to v2.5.0. This is over a year
old and catches us up quite far.

There was one manual change in the cloud_event_controller_test.go file. CloudEvents
changed the string representation of an event, which our tests looked for. This
matches the change to their upstream test: github.com/cloudevents/sdk-go/pull/639.

Signed-off-by: Dan Lorenc <lorenc.d@gmail.com>
Signed-off-by: Dan Lorenc <dlorenc@google.com>