Support activeDeadlineSeconds for Tekton pods 🦌
Kubernetes (and OpenShift) mark `Pod` as either Terminating — has a
"relatively" short life and will terminate at some point — and
NonTerminating — is supposed to run for ever. Kubernetes does the
difference between the two using the `ActiveDeadlineSeconds` field. A
`Pod` with `activeDeadlineSeconds` set will be considered as
Terminating. For example, `Job`'s `Pod` have this field set and are
considered as Terminated.

Currently the pods created by tekton fall under the NonTerminating
quota limits of Kubernetes and OpenShift. This can create issues as
generally builds should fall under the separate terminating quota
limits.

This sets the `activeDeadlineSeconds` field or TaskRun's `Pod` so that
they are considered Terminating. It also sets the value of this field
to a higher value than the specified (or default) Timeout set on the
TaskRun so that Kubernetes won't try to terminate the `Pod` before
Pipeline does.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>